Remove @Deprecated StubPackageManager.
Remove both implementations, push all non-implemented methods down to
DefaultPackageManager.
