--- robolectric/src/main/java/org/robolectric/res/builder/StubPackageManager.java
@@ -1,13 +1,0 @@
-package org.robolectric.res.builder;
-
-/**
- * @deprecated Use {@link org.robolectric.shadows.ShadowPackageManager} instead.
- * <pre>
- *   ShadowPackageManager shadowPackageManager = shadowOf(context.getPackageManager());
- * </pre>
- *
- * If there is functionality you are missing you can extend ShadowPackageManager.
- */
-@Deprecated
-public class StubPackageManager extends org.robolectric.android.StubPackageManager {
-}
